requirements.txt
flask==3.0.0
twilio==8.11.0
chromadb==0.4.22
groq==0.4.2
python-dotenv==1.0.0
requests==2.31.0



